Rename Cardano.CLI.EraBased.Certificate module to Cardano.CLI.EraBased.Run.Certificate
