+++ cardano-cli/src/Cardano/CLI/EraBased/Run/Certificate.hs
@@ -4,7 +4,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 
-module Cardano.CLI.EraBased.Certificate
+module Cardano.CLI.EraBased.Run.Certificate
   ( EraBasedDelegationError(..)
   , runGovernanceDelegrationCertificate
